A bunch of fixes
The fixes includes:
- Unused imports removal
- Lambda becomes defs
- f-strings removal
- `__init__.py` rewrite
- `await ctx.typing()` uses `async with` now to prevent infinite typing
- Some small reformat I think.
- Some names change.
This is a very big commit lol. I think there's more tho..
