+++ embedcreator/embedcreator.py
@@ -10,7 +10,7 @@ from discord import Embed
 from redbot.core import commands
 from redbot.core.bot import Red
 from redbot.core.commands import BadArgument, ColourConverter, Context, FlagConverter
-from redbot.core.utils.chat_formatting import bold, box, text_to_file
+from redbot.core.utils.chat_formatting import box, text_to_file
 
 JSON_EMOJI = discord.PartialEmoji(name="json", animated=False, id=1110250547254141049)
 
@@ -387,7 +387,9 @@ class EmbedEditorView(discord.ui.View):
         self.message: Optional[discord.Message] = None
 
     @discord.ui.button(label="Title", style=discord.ButtonStyle.grey)
-    async def edit_title_button(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def edit_title_button(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
         await interaction.response.send_modal(EmbedTitleModal(self))
 
     @discord.ui.button(label="Description", style=discord.ButtonStyle.grey)
@@ -406,14 +408,20 @@ class EmbedEditorView(discord.ui.View):
     async def edit_colour_button(
         self, interaction: discord.Interaction, button: discord.ui.Button
     ):
-        await interaction.response.send_modal(EmbedColourModal(self, context=self.context))
+        await interaction.response.send_modal(
+            EmbedColourModal(self, context=self.context)
+        )
 
     @discord.ui.button(label="URL", style=discord.ButtonStyle.grey)
-    async def edit_url_button(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def edit_url_button(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
         await interaction.response.send_modal(EmbedURLModal(self))
 
     @discord.ui.button(label="Image", row=1, style=discord.ButtonStyle.grey)
-    async def edit_image_button(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def edit_image_button(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
         await interaction.response.send_modal(EmbedImageModal(self))
 
     @discord.ui.button(label="Thumbnail", row=1, style=discord.ButtonStyle.grey)
@@ -435,7 +443,9 @@ class EmbedEditorView(discord.ui.View):
         await interaction.response.send_modal(EmbedFooterBuilder(self))
 
     @discord.ui.button(label="Clear", row=1, style=discord.ButtonStyle.red)
-    async def clear_button(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def clear_button(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
         embed = discord.Embed(description="[…]")
         self.embed = embed
         self.content = None
@@ -444,7 +454,9 @@ class EmbedEditorView(discord.ui.View):
     @discord.ui.button(
         label="Add field", style=discord.ButtonStyle.green, emoji="\U00002795", row=2
     )
-    async def add_field_button(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def add_field_button(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
         await interaction.response.send_modal(EmbedFieldAdder(self))
 
     @discord.ui.button(
@@ -465,7 +477,9 @@ class EmbedEditorView(discord.ui.View):
     @discord.ui.button(
         label="Get Python", style=discord.ButtonStyle.blurple, emoji="\U0001f40d", row=3
     )
-    async def get_python(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def get_python(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
         embed = self.embed
         text = "embed = discord.Embed("
         if embed.title:
@@ -494,7 +508,9 @@ class EmbedEditorView(discord.ui.View):
                 if gattr := getattr(embed.author, attr):
                     attrs[attr] = gattr
             text += (
-                f"embed.set_author(" + ", ".join(f"{k}={v!r}" for k, v in attrs.items()) + ")\n"
+                "embed.set_author("
+                + ", ".join(f"{k}={v!r}" for k, v in attrs.items())
+                + ")\n"
             )
 
         if embed.footer:
@@ -503,7 +519,9 @@ class EmbedEditorView(discord.ui.View):
                 if gattr := getattr(embed.footer, attr):
                     attrs[attr] = gattr
             text += (
-                f"embed.set_footer(" + ", ".join(f"{k}={v!r}" for k, v in attrs.items()) + ")\n"
+                "embed.set_footer("
+                + ", ".join(f"{k}={v!r}" for k, v in attrs.items())
+                + ")\n"
             )
 
         if not text.endswith("\n\n"):
@@ -536,7 +554,9 @@ class EmbedEditorView(discord.ui.View):
     @discord.ui.button(
         label="Get JSON", style=discord.ButtonStyle.blurple, emoji=JSON_EMOJI, row=3
     )
-    async def get_json(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def get_json(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
         text = json.dumps(self.embed.to_dict(), indent=4)
         if len(text) > 1990:
             await interaction.response.send_message(
@@ -555,8 +575,12 @@ class EmbedEditorView(discord.ui.View):
         emoji=JSON_EMOJI,
         row=3,
     )
-    async def replace_json(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await interaction.response.send_modal(EmbedDictionaryUpdater(self, replace=True))
+    async def replace_json(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
+        await interaction.response.send_modal(
+            EmbedDictionaryUpdater(self, replace=True)
+        )
 
     @discord.ui.button(
         label="Update JSON",
@@ -564,8 +588,12 @@ class EmbedEditorView(discord.ui.View):
         emoji=JSON_EMOJI,
         row=3,
     )
-    async def update_json(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await interaction.response.send_modal(EmbedDictionaryUpdater(self, replace=False))
+    async def update_json(
+        self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
+        await interaction.response.send_modal(
+            EmbedDictionaryUpdater(self, replace=False)
+        )
 
     @discord.ui.select(
         cls=discord.ui.ChannelSelect,
@@ -580,15 +608,19 @@ class EmbedEditorView(discord.ui.View):
         placeholder="Send your embed",
         row=4,
     )
-    async def send(self, interaction: discord.Interaction, select: discord.ui.ChannelSelect):
+    async def send(
+        self, interaction: discord.Interaction, select: discord.ui.ChannelSelect
+    ):
         channel = interaction.guild.get_channel(select.values[0].id)
         if not channel.permissions_for(interaction.guild.me).send_messages:
             return await interaction.response.send_message(
-                f"I do not have permissions to post in {channel.mention}.", ephemeral=True
+                f"I do not have permissions to post in {channel.mention}.",
+                ephemeral=True,
             )
         if not channel.permissions_for(interaction.user).send_messages:
             return await interaction.response.send_message(
-                f"You do not have permissions to post in {channel.mention}.", ephemeral=True
+                f"You do not have permissions to post in {channel.mention}.",
+                ephemeral=True,
             )
         try:
             await channel.send(
@@ -788,7 +820,9 @@ class EmbedCreator(commands.Cog):
         try:
             if options.builder:
                 view.embed = embed
-                view.message = await ctx.send(view=view, embed=embed, content=options.content)
+                view.message = await ctx.send(
+                    view=view, embed=embed, content=options.content
+                )
             else:
                 await ctx.send(embed=embed, content=options.content)
         except discord.HTTPException as exc:
