+++ lock/lock.py
@@ -17,7 +17,9 @@ class Lock(commands.Cog):
 
     def __init__(self, bot: Red):
         self.bot = bot
-        self.config = Config.get_conf(self, identifier=36546565165464, force_registration=True)
+        self.config = Config.get_conf(
+            self, identifier=36546565165464, force_registration=True
+        )
 
         self.config.register_guild(moderator=None, everyone=True, ignore=[])
 
@@ -59,13 +61,19 @@ class Lock(commands.Cog):
         await ctx.send(f"{channel.mention} is already in the ignored channels list.")
 
     @lockset.command(name="unignore")
-    async def lockset_unignore(self, ctx: commands.Context, channel: discord.TextChannel):
+    async def lockset_unignore(
+        self, ctx: commands.Context, channel: discord.TextChannel
+    ):
         """Remove channels from the ignored list."""
         if channel.id not in await self.config.guild(ctx.guild).ignore():
-            return await ctx.send(f"{channel.mention} already isn't in the ignored channels list.")
+            return await ctx.send(
+                f"{channel.mention} already isn't in the ignored channels list."
+            )
         async with self.config.guild(ctx.guild).ignore() as ignore:
             ignore.remove(channel.id)
-        await ctx.send(f"{channel.mention} has been removed from the ignored channels list.")
+        await ctx.send(
+            f"{channel.mention} has been removed from the ignored channels list."
+        )
 
     @lockset.command(name="settings")
     async def lockset_settings(self, ctx: commands.Context):
@@ -85,12 +93,18 @@ class Lock(commands.Cog):
                     c_text.append(c.mention)
             c_text = humanize_list(c_text)
 
-        embed = discord.Embed(colour=await ctx.embed_colour(), timestamp=datetime.datetime.now())
-        embed.set_author(name=ctx.guild.name, icon_url=ctx.guild.icon.url if ctx.guild.icon else None)
+        embed = discord.Embed(
+            colour=await ctx.embed_colour(), timestamp=datetime.datetime.now()
+        )
+        embed.set_author(
+            name=ctx.guild.name, icon_url=ctx.guild.icon.url if ctx.guild.icon else None
+        )
         embed.title = "**__Lock settings:__**"
         embed.set_footer(text="*required to function properly")
 
-        embed.add_field(name="Role that can type after locking*:", value=mods, inline=False)
+        embed.add_field(
+            name="Role that can type after locking*:", value=mods, inline=False
+        )
         embed.add_field(
             name="Using roles to access servers:*:",
             value=str(not data["everyone"]),
@@ -110,7 +124,9 @@ class Lock(commands.Cog):
         which = await self.config.guild(ctx.guild).everyone()
 
         if not mods:
-            return await ctx.send("Uh oh. Looks like your Admins haven't setup this yet.")
+            return await ctx.send(
+                "Uh oh. Looks like your Admins haven't setup this yet."
+            )
         if which:
             await ctx.channel.set_permissions(
                 ctx.guild.default_role, read_messages=True, send_messages=False
@@ -123,7 +139,9 @@ class Lock(commands.Cog):
         await ctx.send(":lock: Channel locked. Only Moderators can type.")
 
     @lock.command(name="server")
-    async def lock_server(self, ctx: commands.Context, confirmation: typing.Optional[bool]):
+    async def lock_server(
+        self, ctx: commands.Context, confirmation: typing.Optional[bool]
+    ):
         """Lock `@everyone` from sending messages in the entire server."""
         if not confirmation:
             return await ctx.send(
@@ -136,7 +154,9 @@ class Lock(commands.Cog):
             ignore = await self.config.guild(ctx.guild).ignore()
 
             if not mods:
-                return await ctx.send("Uh oh. Looks like your Admins haven't setup this yet.")
+                return await ctx.send(
+                    "Uh oh. Looks like your Admins haven't setup this yet."
+                )
             for channel in ctx.guild.text_channels:
                 if channel.id in ignore:
                     continue
@@ -148,7 +168,9 @@ class Lock(commands.Cog):
                     await channel.set_permissions(
                         ctx.guild.default_role, read_messages=False, send_messages=False
                     )
-                await channel.set_permissions(mods, read_messages=True, send_messages=True)
+                await channel.set_permissions(
+                    mods, read_messages=True, send_messages=True
+                )
         await ctx.send(":lock: Server locked. Only Moderators can type.")
 
     @commands.mod()
@@ -161,7 +183,9 @@ class Lock(commands.Cog):
         which = await self.config.guild(ctx.guild).everyone()
 
         if not mods:
-            return await ctx.send("Uh oh. Looks like your Admins haven't setup this yet.")
+            return await ctx.send(
+                "Uh oh. Looks like your Admins haven't setup this yet."
+            )
         if which:
             await ctx.channel.set_permissions(
                 ctx.guild.default_role, read_messages=True, send_messages=True
@@ -181,7 +205,9 @@ class Lock(commands.Cog):
             ignore = await self.config.guild(ctx.guild).ignore()
 
             if not mods:
-                return await ctx.send("Uh oh. Looks like your Admins haven't setup this yet.")
+                return await ctx.send(
+                    "Uh oh. Looks like your Admins haven't setup this yet."
+                )
             for channel in ctx.guild.text_channels:
                 if channel.id in ignore:
                     continue
