+++ minifier/minifier.py
@@ -1,8 +1,7 @@
-import contextlib
 import io
 
 import discord
-import python_minifier as minifier
+import python_minifier as minifier  # type: ignore
 from redbot.core import commands
 from redbot.core.bot import Red
 
@@ -30,19 +29,21 @@ class Minifier(commands.Cog):
 
         You need to attach a file to this command, and it's extension needs to be `.py`.
         """
-        await ctx.typing()
-        if not ctx.message.attachments:
-            return await ctx.send_help()
-        file = ctx.message.attachments[0]
-        file_name = file.filename.lower()
-        if not file_name.endswith((".py", ".python")):
-            return await ctx.send("Must be a python file.")
-        try:
-            file = await file.read()
-        except UnicodeDecodeError:
-            return await ctx.send("Something went wrong when trying to decode this file.")
-        converted = io.BytesIO(minifier.minify(file).encode(encoding="utf-8"))
-        return await ctx.send(
-            content="Please see the attached file below, with your minified code.",
-            file=discord.File(converted, filename=file_name),
-        )
+        async with ctx.typing():
+            if not ctx.message.attachments:
+                return await ctx.send_help()
+            file = ctx.message.attachments[0]
+            file_name = file.filename.lower()
+            if not file_name.endswith((".py", ".python")):
+                return await ctx.send("Must be a python file.")
+            try:
+                file = await file.read()
+            except UnicodeDecodeError:
+                return await ctx.send(
+                    "Something went wrong when trying to decode this file."
+                )
+            converted = io.BytesIO(minifier.minify(file).encode(encoding="utf-8"))
+            return await ctx.send(
+                content="Please see the attached file below, with your minified code.",
+                file=discord.File(converted, filename=file_name),
+            )
