+++ textfont/__init__.py
@@ -1,1195 +1,9 @@
-import discord
-from redbot.core import commands
 from redbot.core.bot import Red
 from redbot.core.utils import get_end_user_data_statement
 
+from .textfont import TextFont
+
 __red_end_user_data_statement__ = get_end_user_data_statement(__file__)
-
-FONTS = {
-    "bold": {
-        "A": "𝗔",
-        "B": "𝗕",
-        "C": "𝗖",
-        "D": "𝗗",
-        "E": "𝗘",
-        "F": "𝗙",
-        "G": "𝗚",
-        "H": "𝗛",
-        "I": "𝗜",
-        "J": "𝗝",
-        "K": "𝗞",
-        "L": "𝗟",
-        "M": "𝗠",
-        "N": "𝗡",
-        "O": "𝗢",
-        "P": "𝗣",
-        "Q": "𝗤",
-        "R": "𝗥",
-        "S": "𝗦",
-        "T": "𝗧",
-        "U": "𝗨",
-        "V": "𝗩",
-        "W": "𝗪",
-        "X": "𝗫",
-        "Y": "𝗬",
-        "Z": "𝗭",
-        "a": "𝗮",
-        "b": "𝗯",
-        "c": "𝗰",
-        "d": "𝗱",
-        "e": "𝗲",
-        "f": "𝗳",
-        "g": "𝗴",
-        "h": "𝗵",
-        "i": "𝗶",
-        "j": "𝗷",
-        "k": "𝗸",
-        "l": "𝗹",
-        "m": "𝗺",
-        "n": "𝗻",
-        "o": "𝗼",
-        "p": "𝗽",
-        "q": "𝗾",
-        "r": "𝗿",
-        "s": "𝘀",
-        "t": "𝘁",
-        "u": "𝘂",
-        "v": "𝘃",
-        "w": "𝘄",
-        "x": "𝘅",
-        "y": "𝘆",
-        "z": "𝘇",
-        "0": "𝟎",
-        "1": "𝟏",
-        "2": "𝟐",
-        "3": "𝟑",
-        "4": "𝟒",
-        "5": "𝟓",
-        "6": "𝟖",
-        "7": "𝟕",
-        "8": "𝟖",
-        "9": "𝟗",
-    },
-    "italic": {
-        "A": "𝘈",
-        "B": "𝘉",
-        "C": "𝘊",
-        "D": "𝘋",
-        "E": "𝘌",
-        "F": "𝘍",
-        "G": "𝘎",
-        "H": "𝘏",
-        "I": "𝘐",
-        "J": "𝘑",
-        "K": "𝘒",
-        "L": "𝘓",
-        "M": "𝘔",
-        "N": "𝘕",
-        "O": "𝘖",
-        "P": "𝘗",
-        "Q": "𝘘",
-        "R": "𝘙",
-        "S": "𝘚",
-        "T": "𝘛",
-        "U": "𝘜",
-        "V": "𝘝",
-        "W": "𝘞",
-        "X": "𝘟",
-        "Y": "𝘠",
-        "Z": "𝘡",
-        "a": "𝘢",
-        "b": "𝘣",
-        "c": "𝘤",
-        "d": "𝘥",
-        "e": "𝘦",
-        "f": "𝘧",
-        "g": "𝘨",
-        "h": "𝘩",
-        "i": "𝘪",
-        "j": "𝘫",
-        "k": "𝘬",
-        "l": "𝘭",
-        "m": "𝘮",
-        "n": "𝘯",
-        "o": "𝘰",
-        "p": "𝘱",
-        "q": "𝘲",
-        "r": "𝘳",
-        "s": "𝘴",
-        "t": "𝘵",
-        "u": "𝘶",
-        "v": "𝘷",
-        "w": "𝘸",
-        "x": "𝘹",
-        "y": "𝘺",
-        "z": "𝘻",
-        "0": "𝟬",
-        "1": "𝟭",
-        "2": "𝟮",
-        "3": "𝟯",
-        "4": "𝟰",
-        "5": "𝟱",
-        "6": "𝟲",
-        "7": "𝟳",
-        "8": "𝟴",
-        "9": "𝟵",
-    },
-    "bold_italic": {
-        "A": "𝘼",
-        "B": "𝘽",
-        "C": "𝘾",
-        "D": "𝘿",
-        "E": "𝑬",
-        "F": "𝑭",
-        "G": "𝑮",
-        "H": "𝐻",
-        "I": "𝐼",
-        "J": "𝑱",
-        "K": "𝑲",
-        "L": "𝐿",
-        "M": "𝑴",
-        "N": "𝑵",
-        "O": "𝑶",
-        "P": "𝑷",
-        "Q": "𝑸",
-        "R": "𝑹",
-        "S": "𝑺",
-        "T": "𝑻",
-        "U": "𝑼",
-        "V": "𝑽",
-        "W": "𝑾",
-        "X": "𝑿",
-        "Y": "𝒀",
-        "Z": "𝒁",
-        "a": "𝒂",
-        "b": "𝒃",
-        "c": "𝒄",
-        "d": "𝒅",
-        "e": "𝒆",
-        "f": "𝒇",
-        "g": "𝒈",
-        "h": "𝒉",
-        "i": "𝒊",
-        "j": "𝒋",
-        "k": "𝒌",
-        "l": "𝒍",
-        "m": "𝒎",
-        "n": "𝒏",
-        "o": "𝒐",
-        "p": "𝒑",
-        "q": "𝒒",
-        "r": "𝒓",
-        "s": "𝒔",
-        "t": "𝒕",
-        "u": "𝒖",
-        "v": "𝒗",
-        "w": "𝒘",
-        "x": "𝒙",
-        "y": "𝒚",
-        "z": "𝒛",
-        "0": "𝟬",
-        "1": "𝟭",
-        "2": "𝟮",
-        "3": "𝟯",
-        "4": "𝟰",
-        "5": "𝟱",
-        "6": "𝟲",
-        "7": "𝟳",
-        "8": "𝟴",
-        "9": "𝟵",
-    },
-    "script": {
-        "A": "𝒜",
-        "B": "𝐵",
-        "C": "𝒞",
-        "D": "𝒟",
-        "E": "𝐸",
-        "F": "𝐹",
-        "G": "𝒢",
-        "H": "𝐻",
-        "I": "𝐼",
-        "J": "𝒥",
-        "K": "𝒦",
-        "L": "𝐿",
-        "M": "𝒲",
-        "N": "𝒩",
-        "O": "𝒪",
-        "P": "𝒫",
-        "Q": "𝒬",
-        "R": "𝒜",
-        "S": "𝒮",
-        "T": "𝒯",
-        "U": "𝒰",
-        "V": "𝒱",
-        "W": "𝒲",
-        "X": "𝒳",
-        "Y": "𝒴",
-        "Z": "𝒵",
-        "a": "𝒶",
-        "b": "𝒷",
-        "c": "𝒸",
-        "d": "𝒹",
-        "e": "𝑒",
-        "f": "𝒻",
-        "g": "𝑔",
-        "h": "𝒽",
-        "i": "𝒾",
-        "j": "𝒿",
-        "k": "𝒿",
-        "l": "𝓁",
-        "m": "𝓂",
-        "n": "𝓃",
-        "o": "𝑜",
-        "p": "𝓅",
-        "q": "𝓆",
-        "r": "𝓇",
-        "s": "𝓈",
-        "t": "𝓉",
-        "u": "𝓊",
-        "v": "𝓋",
-        "w": "𝓌",
-        "x": "𝓍",
-        "y": "𝓎",
-        "z": "𝓏",
-        "0": "𝟬",
-        "1": "𝟭",
-        "2": "𝟮",
-        "3": "𝟯",
-        "4": "𝟰",
-        "5": "𝟱",
-        "6": "𝟲",
-        "7": "𝟳",
-        "8": "𝟴",
-        "9": "𝟵",
-    },
-    "sans_serif": {
-        "A": "𝖠",
-        "B": "𝖡",
-        "C": "𝖢",
-        "D": "𝖣",
-        "E": "𝖤",
-        "F": "𝖥",
-        "G": "𝖦",
-        "H": "𝖧",
-        "I": "𝖨",
-        "J": "𝖩",
-        "K": "𝖪",
-        "L": "𝖫",
-        "M": "𝖬",
-        "N": "𝖭",
-        "O": "𝖮",
-        "P": "𝖯",
-        "Q": "𝖰",
-        "R": "𝖱",
-        "S": "𝖲",
-        "T": "𝖳",
-        "U": "𝖴",
-        "V": "𝖵",
-        "W": "𝖶",
-        "X": "𝖷",
-        "Y": "𝖸",
-        "Z": "𝖹",
-        "a": "𝖺",
-        "b": "𝖻",
-        "c": "𝖼",
-        "d": "𝖽",
-        "e": "𝖾",
-        "f": "𝖿",
-        "g": "𝗀",
-        "h": "𝗁",
-        "i": "𝗂",
-        "j": "𝗃",
-        "k": "𝗄",
-        "l": "𝗅",
-        "m": "𝗆",
-        "n": "𝗇",
-        "o": "𝗈",
-        "p": "𝗉",
-        "q": "𝗊",
-        "r": "𝗋",
-        "s": "𝗌",
-        "t": "𝗍",
-        "u": "𝗎",
-        "v": "𝗏",
-        "w": "𝗐",
-        "x": "𝗑",
-        "y": "𝗒",
-        "z": "𝗓",
-        "0": "𝟢",
-        "1": "𝟣",
-        "2": "𝟤",
-        "3": "𝟥",
-        "4": "𝟦",
-        "5": "𝟧",
-        "6": "𝟨",
-        "7": "𝟩",
-        "8": "𝟪",
-        "9": "𝟫",
-    },
-    "sans_serif_bold": {
-        "A": "𝗔",
-        "B": "𝗕",
-        "C": "𝗖",
-        "D": "𝗗",
-        "E": "𝗘",
-        "F": "𝗙",
-        "G": "𝗚",
-        "H": "𝗛",
-        "I": "𝗜",
-        "J": "𝗝",
-        "K": "𝗞",
-        "L": "𝗟",
-        "M": "𝗠",
-        "N": "𝗡",
-        "O": "𝗢",
-        "P": "𝗣",
-        "Q": "𝗤",
-        "R": "𝗥",
-        "S": "𝗦",
-        "T": "𝗧",
-        "U": "𝗨",
-        "V": "𝗩",
-        "W": "𝗪",
-        "X": "𝗫",
-        "Y": "𝗬",
-        "Z": "𝗭",
-        "a": "𝗮",
-        "b": "𝗯",
-        "c": "𝗰",
-        "d": "𝗱",
-        "e": "𝗲",
-        "f": "𝗳",
-        "g": "𝗴",
-        "h": "𝗵",
-        "i": "𝗶",
-        "j": "𝗷",
-        "k": "𝗸",
-        "l": "𝗹",
-        "m": "𝗺",
-        "n": "𝗻",
-        "o": "𝗼",
-        "p": "𝗽",
-        "q": "𝗾",
-        "r": "𝗿",
-        "s": "𝘀",
-        "t": "𝘁",
-        "u": "𝘂",
-        "v": "𝘃",
-        "w": "𝘄",
-        "x": "𝘅",
-        "y": "𝘆",
-        "z": "𝘇",
-        "0": "𝟬",
-        "1": "𝟭",
-        "2": "𝟮",
-        "3": "𝟯",
-        "4": "𝟰",
-        "5": "𝟱",
-        "6": "𝟲",
-        "7": "𝟳",
-        "8": "𝟴",
-        "9": "𝟵",
-    },
-    "sans_serif_bold_italic": {
-        "A": "𝘼",
-        "B": "𝘽",
-        "C": "𝘾",
-        "D": "𝘿",
-        "E": "𝙀",
-        "F": "𝙁",
-        "G": "𝙂",
-        "H": "𝙃",
-        "I": "𝙄",
-        "J": "𝙅",
-        "K": "𝙆",
-        "L": "𝙇",
-        "M": "𝙈",
-        "N": "𝙉",
-        "O": "𝙊",
-        "P": "𝙋",
-        "Q": "𝙌",
-        "R": "𝙍",
-        "S": "𝙎",
-        "T": "𝙏",
-        "U": "𝙐",
-        "V": "𝙑",
-        "W": "𝙒",
-        "X": "𝙓",
-        "Y": "𝙔",
-        "Z": "𝙕",
-        "a": "𝙖",
-        "b": "𝙗",
-        "c": "𝙘",
-        "d": "𝙙",
-        "e": "𝙚",
-        "f": "𝙛",
-        "g": "𝙜",
-        "h": "𝙝",
-        "i": "𝙞",
-        "j": "𝙟",
-        "k": "𝙠",
-        "l": "𝙡",
-        "m": "𝙢",
-        "n": "𝙣",
-        "o": "𝙤",
-        "p": "𝙥",
-        "q": "𝙦",
-        "r": "𝙧",
-        "s": "𝙨",
-        "t": "𝙩",
-        "u": "𝙪",
-        "v": "𝙫",
-        "w": "𝙬",
-        "x": "𝙭",
-        "y": "𝙮",
-        "z": "𝙯",
-        "0": "𝟬",
-        "1": "𝟭",
-        "2": "𝟮",
-        "3": "𝟯",
-        "4": "𝟰",
-        "5": "𝟱",
-        "6": "𝟲",
-        "7": "𝟳",
-        "8": "𝟴",
-        "9": "𝟵",
-    },
-    "sans_serif_italic": {
-        "A": "𝘈",
-        "B": "𝘉",
-        "C": "𝘊",
-        "D": "𝘋",
-        "E": "𝘌",
-        "F": "𝘍",
-        "G": "𝘎",
-        "H": "𝘏",
-        "I": "𝘐",
-        "J": "𝘑",
-        "K": "𝘒",
-        "L": "𝘓",
-        "M": "𝘔",
-        "N": "𝘕",
-        "O": "𝘖",
-        "P": "𝘗",
-        "Q": "𝘘",
-        "R": "𝘙",
-        "S": "𝘚",
-        "T": "𝘛",
-        "U": "𝘜",
-        "V": "𝘝",
-        "W": "𝘞",
-        "X": "𝘟",
-        "Y": "𝘠",
-        "Z": "𝘡",
-        "a": "𝘢",
-        "b": "𝘣",
-        "c": "𝘤",
-        "d": "𝘥",
-        "e": "𝘦",
-        "f": "𝘧",
-        "g": "𝘨",
-        "h": "𝘩",
-        "i": "𝘪",
-        "j": "𝘫",
-        "k": "𝘬",
-        "l": "𝘭",
-        "m": "𝘮",
-        "n": "𝘯",
-        "o": "𝘰",
-        "p": "𝘱",
-        "q": "𝘲",
-        "r": "𝘳",
-        "s": "𝘴",
-        "t": "𝘵",
-        "u": "𝘶",
-        "v": "𝘷",
-        "w": "𝘸",
-        "x": "𝘹",
-        "y": "𝘺",
-        "z": "𝘻",
-    },
-    "monospace": {
-        "A": "𝙰",
-        "B": "𝙱",
-        "C": "𝙲",
-        "D": "𝙳",
-        "E": "𝙴",
-        "F": "𝙵",
-        "G": "𝙶",
-        "H": "𝙷",
-        "I": "𝙸",
-        "J": "𝙹",
-        "K": "𝙺",
-        "L": "𝙻",
-        "M": "𝙼",
-        "N": "𝙽",
-        "O": "𝙾",
-        "P": "𝙿",
-        "Q": "𝚀",
-        "R": "𝚁",
-        "S": "𝚂",
-        "T": "𝚃",
-        "U": "𝚄",
-        "V": "𝚅",
-        "W": "𝚆",
-        "X": "𝚇",
-        "Y": "𝚈",
-        "Z": "𝚉",
-        "a": "𝚊",
-        "b": "𝚋",
-        "c": "𝚌",
-        "d": "𝚍",
-        "e": "𝚎",
-        "f": "𝚏",
-        "g": "𝚐",
-        "h": "𝚑",
-        "i": "𝚒",
-        "j": "𝚓",
-        "k": "𝚔",
-        "l": "𝚕",
-        "m": "𝚖",
-        "n": "𝚗",
-        "o": "𝚘",
-        "p": "𝚙",
-        "q": "𝚚",
-        "r": "𝚛",
-        "s": "𝚜",
-        "t": "𝚝",
-        "u": "𝚞",
-        "v": "𝚟",
-        "w": "𝚠",
-        "x": "𝚡",
-        "y": "𝚢",
-        "z": "𝚣",
-        "0": "𝟶",
-        "1": "𝟷",
-        "2": "𝟸",
-        "3": "𝟹",
-        "4": "𝟺",
-        "5": "𝟻",
-        "6": "𝟼",
-        "7": "𝟽",
-        "8": "𝟾",
-        "9": "𝟿",
-    },
-    "fraktur": {
-        "A": "𝔄",
-        "B": "𝔅",
-        "C": "ℭ",
-        "D": "𝔇",
-        "E": "𝔈",
-        "F": "𝔉",
-        "G": "𝔊",
-        "H": "ℌ",
-        "I": "ℑ",
-        "J": "𝔍",
-        "K": "𝔎",
-        "L": "𝔏",
-        "M": "𝔐",
-        "N": "𝔑",
-        "O": "𝔒",
-        "P": "𝔓",
-        "Q": "𝔔",
-        "R": "ℜ",
-        "S": "𝔖",
-        "T": "𝔗",
-        "U": "𝔘",
-        "V": "𝔙",
-        "W": "𝔚",
-        "X": "𝔛",
-        "Y": "𝔜",
-        "Z": "ℨ",
-        "a": "𝔞",
-        "b": "𝔟",
-        "c": "𝔠",
-        "d": "𝔡",
-        "e": "𝔢",
-        "f": "𝔣",
-        "g": "𝔤",
-        "h": "𝔥",
-        "i": "𝔦",
-        "j": "𝔧",
-        "k": "𝔨",
-        "l": "𝔩",
-        "m": "𝔪",
-        "n": "𝔫",
-        "o": "𝔬",
-        "p": "𝔭",
-        "q": "𝔮",
-        "r": "𝔯",
-        "s": "𝔰",
-        "t": "𝔱",
-        "u": "𝔲",
-        "v": "𝔳",
-        "w": "𝔴",
-        "x": "𝔵",
-        "y": "𝔶",
-        "z": "𝔷",
-    },
-    "double_struck": {
-        "A": "𝔸",
-        "B": "𝔹",
-        "C": "ℂ",
-        "D": "𝔻",
-        "E": "𝔼",
-        "F": "𝔽",
-        "G": "𝔾",
-        "H": "ℍ",
-        "I": "𝕀",
-        "J": "𝕁",
-        "K": "𝕂",
-        "L": "𝕃",
-        "M": "𝕄",
-        "N": "ℕ",
-        "O": "𝕆",
-        "P": "ℙ",
-        "Q": "ℚ",
-        "R": "ℝ",
-        "S": "𝕊",
-        "T": "𝕋",
-        "U": "𝕌",
-        "V": "𝕍",
-        "W": "𝕎",
-        "X": "𝕏",
-        "Y": "𝕐",
-        "Z": "ℤ",
-        "a": "𝕒",
-        "b": "𝕓",
-        "c": "𝕔",
-        "d": "𝕕",
-        "e": "𝕖",
-        "f": "𝕗",
-        "g": "𝕘",
-        "h": "𝕙",
-        "i": "𝕚",
-        "j": "𝕛",
-        "k": "𝕜",
-        "l": "𝕝",
-        "m": "𝕞",
-        "n": "𝕟",
-        "o": "𝕠",
-        "p": "𝕡",
-        "q": "𝕢",
-        "r": "𝕣",
-        "s": "𝕤",
-        "t": "𝕥",
-        "u": "𝕦",
-        "v": "𝕧",
-        "w": "𝕨",
-        "x": "𝕩",
-        "y": "𝕪",
-        "z": "𝕫",
-        "0": "𝟘",
-        "1": "𝟙",
-        "2": "𝟚",
-        "3": "𝟛",
-        "4": "𝟜",
-        "5": "𝟝",
-        "6": "𝟞",
-        "7": "𝟟",
-        "8": "𝟠",
-        "9": "𝟡",
-    },
-    "circled": {
-        "A": "Ⓐ",
-        "B": "Ⓑ",
-        "C": "Ⓒ",
-        "D": "Ⓓ",
-        "E": "Ⓔ",
-        "F": "Ⓕ",
-        "G": "Ⓖ",
-        "H": "Ⓗ",
-        "I": "Ⓘ",
-        "J": "Ⓙ",
-        "K": "Ⓚ",
-        "L": "Ⓛ",
-        "M": "Ⓜ",
-        "N": "Ⓝ",
-        "O": "Ⓞ",
-        "P": "Ⓟ",
-        "Q": "Ⓠ",
-        "R": "Ⓡ",
-        "S": "Ⓢ",
-        "T": "Ⓣ",
-        "U": "Ⓤ",
-        "V": "Ⓥ",
-        "W": "Ⓦ",
-        "X": "Ⓧ",
-        "Y": "Ⓨ",
-        "Z": "Ⓩ",
-        "a": "ⓐ",
-        "b": "ⓑ",
-        "c": "ⓒ",
-        "d": "ⓓ",
-        "e": "ⓔ",
-        "f": "ⓕ",
-        "g": "ⓖ",
-        "h": "ⓗ",
-        "i": "ⓘ",
-        "j": "ⓙ",
-        "k": "ⓚ",
-        "l": "ⓛ",
-        "m": "ⓜ",
-        "n": "ⓝ",
-        "o": "ⓞ",
-        "p": "ⓟ",
-        "q": "ⓠ",
-        "r": "ⓡ",
-        "s": "ⓢ",
-        "t": "ⓣ",
-        "u": "ⓤ",
-        "v": "ⓥ",
-        "w": "ⓦ",
-        "x": "ⓧ",
-        "y": "ⓨ",
-        "z": "ⓩ",
-        "0": "⓪",
-        "1": "①",
-        "2": "②",
-        "3": "③",
-        "4": "④",
-        "5": "⑤",
-        "6": "⑥",
-        "7": "⑦",
-        "8": "⑧",
-        "9": "⑨",
-    },
-    "circled_fill": {
-        "A": "🅐",
-        "B": "🅑",
-        "C": "🅒",
-        "D": "🅓",
-        "E": "🅔",
-        "F": "🅕",
-        "G": "🅖",
-        "H": "🅗",
-        "I": "🅘",
-        "J": "🅙",
-        "K": "🅚",
-        "L": "🅛",
-        "M": "🅜",
-        "N": "🅝",
-        "O": "🅞",
-        "P": "🅟",
-        "Q": "🅠",
-        "R": "🅡",
-        "S": "🅢",
-        "T": "🅣",
-        "U": "🅤",
-        "V": "🅥",
-        "W": "🅦",
-        "X": "🅧",
-        "Y": "🅨",
-        "Z": "🅩",
-        "a": "🅐",
-        "b": "🅑",
-        "c": "🅒",
-        "d": "🅓",
-        "e": "🅔",
-        "f": "🅕",
-        "g": "🅖",
-        "h": "🅗",
-        "i": "🅘",
-        "j": "🅙",
-        "k": "🅚",
-        "l": "🅛",
-        "m": "🅜",
-        "n": "🅝",
-        "o": "🅞",
-        "p": "🅟",
-        "q": "🅠",
-        "r": "🅡",
-        "s": "🅢",
-        "t": "🅣",
-        "u": "🅤",
-        "v": "🅥",
-        "w": "🅦",
-        "x": "🅧",
-        "y": "🅨",
-        "z": "🅩",
-    },
-    "small_caps": {
-        "A": "ᴀ",
-        "B": "ʙ",
-        "C": "ᴄ",
-        "D": "ᴅ",
-        "E": "ᴇ",
-        "F": "ғ",
-        "G": "ɢ",
-        "H": "ʜ",
-        "I": "ɪ",
-        "J": "ᴊ",
-        "K": "ᴋ",
-        "L": "ʟ",
-        "M": "ᴍ",
-        "N": "ɴ",
-        "O": "ᴏ",
-        "P": "ᴘ",
-        "Q": "ǫ",
-        "R": "ʀ",
-        "S": "s",
-        "T": "ᴛ",
-        "U": "ᴜ",
-        "V": "ᴠ",
-        "W": "ᴡ",
-        "X": "x",
-        "Y": "ʏ",
-        "Z": "ᴢ",
-        "a": "ᴀ",
-        "b": "ʙ",
-        "c": "ᴄ",
-        "d": "ᴅ",
-        "e": "ᴇ",
-        "f": "ғ",
-        "g": "ɢ",
-        "h": "ʜ",
-        "i": "ɪ",
-        "j": "ᴊ",
-        "k": "ᴋ",
-        "l": "ʟ",
-        "m": "ᴍ",
-        "n": "ɴ",
-        "o": "ᴏ",
-        "p": "ᴘ",
-        "q": "ǫ",
-        "r": "ʀ",
-        "s": "s",
-        "t": "ᴛ",
-        "u": "ᴜ",
-        "v": "ᴠ",
-        "w": "ᴡ",
-        "x": "x",
-        "y": "ʏ",
-        "z": "ᴢ",
-    },
-    "parenthesized": {
-        "A": "🄐",
-        "B": "🄑",
-        "C": "🄒",
-        "D": "🄓",
-        "E": "🄔",
-        "F": "🄕",
-        "G": "🄖",
-        "H": "🄗",
-        "I": "🄘",
-        "J": "🄙",
-        "K": "🄚",
-        "L": "🄛",
-        "M": "🄜",
-        "N": "🄝",
-        "O": "🄞",
-        "P": "🄟",
-        "Q": "🄠",
-        "R": "🄡",
-        "S": "🄢",
-        "T": "🄣",
-        "U": "🄤",
-        "V": "🄥",
-        "W": "🄦",
-        "X": "🄧",
-        "Y": "🄨",
-        "Z": "🄩",
-        "a": "🄐",
-        "b": "🄑",
-        "c": "🄒",
-        "d": "🄓",
-        "e": "🄔",
-        "f": "🄕",
-        "g": "🄖",
-        "h": "🄗",
-        "i": "🄘",
-        "j": "🄙",
-        "k": "🄚",
-        "l": "🄛",
-        "m": "🄜",
-        "n": "🄝",
-        "o": "🄞",
-        "p": "🄟",
-        "q": "🄠",
-        "r": "🄡",
-        "s": "🄢",
-        "t": "🄣",
-        "u": "🄤",
-        "v": "🄥",
-        "w": "🄦",
-        "x": "🄧",
-        "y": "🄨",
-        "z": "🄩",
-        "0": "⑴",
-        "1": "⑴",
-        "2": "⑵",
-        "3": "⑶",
-        "4": "⑷",
-        "5": "⑸",
-        "6": "⑹",
-        "7": "⑺",
-        "8": "⑻",
-        "9": "⑼",
-    },
-    "full_width": {
-        "A": "Ａ",
-        "B": "Ｂ",
-        "C": "Ｃ",
-        "D": "Ｄ",
-        "E": "Ｅ",
-        "F": "Ｆ",
-        "G": "Ｇ",
-        "H": "Ｈ",
-        "I": "Ｉ",
-        "J": "Ｊ",
-        "K": "Ｋ",
-        "L": "Ｌ",
-        "M": "Ｍ",
-        "N": "Ｎ",
-        "O": "Ｏ",
-        "P": "Ｐ",
-        "Q": "Ｑ",
-        "R": "Ｒ",
-        "S": "Ｓ",
-        "T": "Ｔ",
-        "U": "Ｕ",
-        "V": "Ｖ",
-        "W": "Ｗ",
-        "X": "Ｘ",
-        "Y": "Ｙ",
-        "Z": "Ｚ",
-        "a": "ａ",
-        "b": "ｂ",
-        "c": "ｃ",
-        "d": "ｄ",
-        "e": "ｅ",
-        "f": "ｆ",
-        "g": "ｇ",
-        "h": "ｈ",
-        "i": "ｉ",
-        "j": "ｊ",
-        "k": "ｋ",
-        "l": "ｌ",
-        "m": "ｍ",
-        "n": "ｎ",
-        "o": "ｏ",
-        "p": "ｐ",
-        "q": "ｑ",
-        "r": "ｒ",
-        "s": "ｓ",
-        "t": "ｔ",
-        "u": "ｕ",
-        "v": "ｖ",
-        "w": "ｗ",
-        "x": "ｘ",
-        "y": "ｙ",
-        "z": "ｚ",
-        "0": "０",
-        "1": "１",
-        "2": "２",
-        "3": "３",
-        "4": "４",
-        "5": "５",
-        "6": "６",
-        "7": "７",
-        "8": "８",
-        "9": "９",
-    },
-    "squared": {
-        "A": "🄰",
-        "B": "🄱",
-        "C": "🄲",
-        "D": "🄳",
-        "E": "🄴",
-        "F": "🄵",
-        "G": "🄶",
-        "H": "🄷",
-        "I": "🄸",
-        "J": "🄹",
-        "K": "🄺",
-        "L": "🄻",
-        "M": "🄼",
-        "N": "🄽",
-        "O": "🄾",
-        "P": "🄿",
-        "Q": "🅀",
-        "R": "🅁",
-        "S": "🅂",
-        "T": "🅃",
-        "U": "🅄",
-        "V": "🅅",
-        "W": "🅆",
-        "X": "🅇",
-        "Y": "🅈",
-        "Z": "🅉",
-        "a": "🄰",
-        "b": "🄱",
-        "c": "🄲",
-        "d": "🄳",
-        "e": "🄴",
-        "f": "🄵",
-        "g": "🄶",
-        "h": "🄷",
-        "i": "🄸",
-        "j": "🄹",
-        "k": "🄺",
-        "l": "🄻",
-        "m": "🄼",
-        "n": "🄽",
-        "o": "🄾",
-        "p": "🄿",
-        "q": "🅀",
-        "r": "🅁",
-        "s": "🅂",
-        "t": "🅃",
-        "u": "🅄",
-        "v": "🅅",
-        "w": "🅆",
-        "x": "🅇",
-        "y": "🅈",
-        "z": "🅉",
-    },
-}
-
-NORMALIZED_FONTS = {
-    "bold": "Bold",
-    "italic": "Italic",
-    "bold_italic": "Bold Italic",
-    "script": "Script",
-    "sans_serif": "Sans Serif",
-    "sans_serif_bold": "Sans Serif Bold",
-    "sans_serif_bold_italic": "Sans Serif Bold Italic",
-    "sans_serif_italic": "Sans Serif Italic",
-    "monospace": "Monospace",
-    "fraktur": "Fraktur",
-    "double_struck": "Double Struck",
-    "circled": "Circled",
-    "circled_fill": "Circled Fill",
-    "small_caps": "Small Caps",
-    "parenthesized": "Parenthesized",
-    "full_width": "Full Width",
-    "squared": "Squared",
-}
-
-
-def render(font: str, text: str):
-    return text.translate(str.maketrans(FONTS[font]))
-
-
-class FontView(discord.ui.View):
-    def __init__(self, text: str):
-        super().__init__()
-        self.text = text
-        self.latest_values: list[str] = []
-        self.add_item(FontSelect(text))
-
-    def format_fonts(self, fonts: list[str] = []):
-        if not fonts:
-            fonts = sorted(FONTS)
-        if len(fonts) == 1:
-            return render(fonts[0], self.text)
-        return "\n".join(
-            f"- {NORMALIZED_FONTS[font]} - {render(font, self.text)}" for font in fonts
-        )
-
-    @discord.ui.button(label="All Fonts", row=2)
-    async def font_select_all_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
-    ):
-        self.latest_values = list(FONTS)
-        await interaction.response.edit_message(content=self.format_fonts())
-
-    @discord.ui.button(label="Change Text", row=2)
-    async def font_change_text_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
-    ):
-        await interaction.response.send_modal(FontTextChangeModal(self))
-
-
-class FontSelect(discord.ui.Select):
-    def __init__(self, text: str):
-        options = [
-            discord.SelectOption(
-                label=NORMALIZED_FONTS[font], description=render(font=font, text=text), value=font
-            )
-            for font in sorted(FONTS)
-        ]
-        super().__init__(
-            placeholder="Choose a font or multiple fonts",
-            max_values=len(options),
-            options=options,
-            row=1,
-        )
-
-        self.view: FontView
-
-    def regenerate_options(self):
-        options = [
-            discord.SelectOption(
-                label=NORMALIZED_FONTS[font],
-                description=render(font=font, text=self.view.text),
-                value=font,
-            )
-            for font in sorted(FONTS)
-        ]
-        self.options = options
-
-    async def callback(self, interaction: discord.Interaction):
-        self.view.latest_values = self.values
-        if len(self.values) > 1:
-            m = "\n".join(
-                f"- {NORMALIZED_FONTS[value]} - {render(value, self.view.text)}"
-                for value in self.values
-            )
-        else:
-            m = render(self.values[0], self.view.text)
-        if len(m) > 2000:
-            return await interaction.response.send_message(
-                "Please do one font at a time, I can't send this many characters!", ephemeral=True
-            )
-        await interaction.response.edit_message(content=m)
-
-
-class FontTextChangeModal(discord.ui.Modal):
-    def __init__(self, view: FontView):
-        super().__init__(title="Change text")
-        self.view = view
-        self.new_text = discord.ui.TextInput(
-            label="New text",
-            default=view.text,
-            placeholder=view.text,
-            style=discord.TextStyle.long,
-            required=True,
-        )
-        self.add_item(self.new_text)
-
-    async def on_submit(self, interaction: discord.Interaction) -> None:
-        self.view.text = new_text = self.new_text.value
-        self.view.children[-1].regenerate_options()
-        if not (latest := self.view.latest_values):
-            content = f"Text: {new_text}\nUse the dropdown below to change the font!"
-        else:
-            content = self.view.format_fonts(latest)
-        await interaction.response.edit_message(content=content, view=self.view)
-
-
-class TextFont(commands.Cog):
-    """Generate text in different fonts."""
-
-    FONTS = FONTS  # debugging access
-
-    __author__ = "Kreusada"
-    __version__ = "1.0.0"
-
-    def format_help_for_context(self, ctx: commands.Context) -> str:
-        context = super().format_help_for_context(ctx)
-        return f"{context}\n\nAuthor: {self.__author__}\nVersion: {self.__version__}"
-
-    async def red_delete_data_for_user(self, **kwargs):
-        return
-
-    @commands.command()
-    async def write(self, ctx: commands.Context, *, text: str):
-        """Write text in different fonts (alphabetic unicode variations)."""
-        content = f"Text: {text}\nUse the dropdown below to change the font!"
-        if len(content) > 2000:
-            return await ctx.send("Please send text with less characters.")
-        await ctx.send(content=content, view=FontView(text))
 
 
 async def setup(bot: Red):
